apps/api/pyproject.toml
[project]
name = "bomatech-api"
version = "0.1.0"
description = "Bomatech REST API — FastAPI + Supabase"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.8",
    "pydantic-settings>=2.5",
    "supabase>=2.8",
    "python-multipart>=0.0.12",
    "httpx>=0.27",
    "anthropic>=0.39",
    "python-jose[cryptography]>=3.3",
    "bomatech-engines",
    "bomatech-ai",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.6",
    "mypy>=1.11",
]

[tool.uv.sources]
bomatech-engines = { path = "../../packages/engines", editable = true }
bomatech-ai = { path = "../../packages/ai", editable = true }

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]
